requirements.txt
# Web Frameworks
fastapi==0.115.5
uvicorn[standard]==0.34.0
streamlit==1.40.2
pydantic==2.10.3
pydantic-settings==2.6.1

# Vector Database
chromadb==0.5.23

# Embeddings
sentence-transformers==3.3.1

# LLM
llama-cpp-python==0.3.2

# MCP Server
mcp==1.1.2

# Background Tasks
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
